flaskProject/requirements.txt
gunicorn==19.7
python-decouple==3.3
whitenoise==3.3.1
psycopg2==2.8.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-RESTful==0.3.8
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-HTTPAuth==4.0.0
passlib==1.7.2
itsdangerous==1.1.0
pytest==4.6.11
geopy==1.23.0
